Add yaml include directive support
