+++ tesseract_srdf/src/srdf_model.cpp
@@ -51,6 +51,7 @@ TESSERACT_COMMON_IGNORE_WARNINGS_POP
 #include <tesseract_srdf/utils.h>
 #include <tesseract_common/utils.h>
 #include <tesseract_common/yaml_utils.h>
+#include <tesseract_common/yaml_extenstions.h>
 #include <tesseract_common/eigen_serialization.h>
 #include <tesseract_common/resource_locator.h>
 #include <tesseract_common/collision_margin_data.h>
